remove admin view course page and merge into student course page
