+++ src/pages/student/dashboard/courses/view-lesson/index.js
@@ -10,7 +10,20 @@ import CourseProgressCard from 'components/Course/ProgressCard'
 import AdditionalContentCard from 'components/Course/AdditionalContentCard'
 import LessonPlaylist from 'components/Course/LessonPlaylist'
 import { useSelector } from 'react-redux'
-import { USER_TYPE_ENUM } from 'constants/constants'
+import { ADMIN_VERIFIED_ENUM, USER_TYPE_ENUM } from 'constants/constants'
+import { acceptCourseRequest, rejectCourseRequest } from 'services/courses/requests'
+import { showNotification } from 'components/utils'
+import {
+  COURSE_ACCEPT_ERROR,
+  COURSE_ACCEPT_SUCCESS,
+  COURSE_REJECT_ERROR,
+  COURSE_REJECT_SUCCESS,
+  ERROR,
+  SUCCESS,
+} from 'constants/notifications'
+import { APPROVE_COURSE, REJECT_COURSE } from 'constants/text'
+import { Button, Space } from 'antd'
+import { CheckOutlined, CloseOutlined } from '@ant-design/icons'
 
 const StudentCourseLesson = () => {
   const { courseId, lessonId } = useParams()
@@ -23,6 +36,26 @@ const StudentCourseLesson = () => {
   const [percent, setPercent] = useState(0)
   const [currentVideoUrl, setCurrentVideoUrl] = useState('')
   const [comments, setComments] = useState([])
+
+  const approveCourse = async () => {
+    const result = await acceptCourseRequest(courseId)
+    if (result && result.success) {
+      viewCourse()
+      showNotification('success', SUCCESS, COURSE_ACCEPT_SUCCESS)
+    } else {
+      showNotification('error', ERROR, COURSE_ACCEPT_ERROR)
+    }
+  }
+
+  const rejectCourse = async () => {
+    const result = await rejectCourseRequest(courseId)
+    if (result && result.success) {
+      viewCourse()
+      showNotification('success', SUCCESS, COURSE_REJECT_SUCCESS)
+    } else {
+      showNotification('error', ERROR, COURSE_REJECT_ERROR)
+    }
+  }
 
   const viewCourse = async () => {
     const result = await getCourseById(courseId)
@@ -80,12 +113,42 @@ const StudentCourseLesson = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
+  const AdminCourseActions = () => {
+    return (
+      <div className="col-12 col-md-auto col-lg-auto mt-4 mt-md-0 text-center text-md-right">
+        <Space size="large">
+          <Button
+            className="btn btn-success text-white"
+            shape="round"
+            size="large"
+            icon={<CheckOutlined />}
+            disabled={currentCourse.adminVerified === ADMIN_VERIFIED_ENUM.ACCEPTED}
+            onClick={() => approveCourse()}
+          >
+            {APPROVE_COURSE}
+          </Button>
+          <Button
+            className="btn btn-danger text-white"
+            shape="round"
+            size="large"
+            icon={<CloseOutlined />}
+            disabled={currentCourse.adminVerified === ADMIN_VERIFIED_ENUM.REJECTED}
+            onClick={() => rejectCourse()}
+          >
+            {REJECT_COURSE}
+          </Button>
+        </Space>
+      </div>
+    )
+  }
+
   return (
     <div>
-      <div className="row pt-2">
+      <div className="row pt-2 justify-content-md-between">
         <div className="col-12 col-md-3 col-lg-2 mt-4 mt-md-0">
           <BackBtn />
         </div>
+        {user.userType === USER_TYPE_ENUM.ADMIN && <AdminCourseActions />}
       </div>
       <div className="row mt-5">
         <div className="col-12 col-lg-8">
@@ -94,12 +157,14 @@ const StudentCourseLesson = () => {
             currentLesson={currentLesson}
             currentVideoUrl={currentVideoUrl}
           />
-          <LessonComments
-            comments={comments}
-            setComments={setComments}
-            lessonId={lessonId}
-            currentLesson={currentLesson}
-          />
+          {user.userType !== USER_TYPE_ENUM.ADMIN && (
+            <LessonComments
+              comments={comments}
+              setComments={setComments}
+              lessonId={lessonId}
+              currentLesson={currentLesson}
+            />
+          )}
         </div>
         <div className="col-12 col-lg-4">
           {user && user.userType === USER_TYPE_ENUM.STUDENT && (
@@ -114,7 +179,7 @@ const StudentCourseLesson = () => {
             currentCourseContract={currentCourseContract}
             currentCourse={currentCourse}
             currentLesson={currentLesson}
-            isAdmin={false}
+            isAdmin={user.userType === USER_TYPE_ENUM.ADMIN}
           />
         </div>
       </div>
